wait for depositorUpdate in cypress test and create analysisPlan before emitting stepper event
