+++ client/src/components/Wizard/WizardNavigationButtons.vue
@@ -70,6 +70,7 @@
 import Button from "../DesignSystem/Button.vue";
 import GoBackDialog from "./GoBackDialog.vue";
 import {mapState, mapGetters} from "vuex";
+import stepInformation, {depositorSteps, STEP_0600_EPSILON_SET} from "@/data/stepInformation";
 
 export default {
   components: {Button, GoBackDialog},
@@ -79,12 +80,41 @@ export default {
   methods: {
     handleContinue: function () {
       window.scrollTo(0, 0);
-      this.$emit("update:stepperPosition", this.stepperPosition + 1);
+      this.updateUserStep()
     },
     handleBack: function () {
       this.dialogGoBack = false;
       this.$emit("update:stepperPosition", this.stepperPosition - 1);
     },
+    emitStepEvent() {
+      this.$emit("update:stepperPosition", this.stepperPosition + 1)
+    },
+    updateUserStep() {
+      const completedStep = stepInformation[this.getDepositorSetupInfo.userStep].nextStep
+      const completedStepProp = {userStep: completedStep}
+      // Update the user step on the DepositorSetup or the Analysis Plan, depending
+      // where we are in the Wizard
+      if (depositorSteps.includes(completedStep)) {
+        const payload = {objectId: this.getDepositorSetupInfo.objectId, props: completedStepProp}
+        this.$store.dispatch('dataset/updateDepositorSetupInfo', payload).then(() => {
+          // if the step that has just been completed is  STEP_0600_EPSILON_SET, then
+          // create the AnalysisPlan before continuing on to the
+          // Create Statistics wizard step
+          if (completedStep === STEP_0600_EPSILON_SET) {
+            this.$store.dispatch('dataset/createAnalysisPlan', this.datasetInfo.objectId)
+                .then(() => this.emitStepEvent())
+          } else {
+            this.emitStepEvent()
+          }
+        })
+
+      } else {
+        const payload = {objectId: this.analysisPlan.objectId, props: completedStepProp}
+        this.$store.dispatch('dataset/updateAnalysisPlan', payload).then(() => this.emitStepEvent())
+      }
+
+    },
+
 
   },
   data: () => ({
@@ -95,7 +125,8 @@ export default {
   }),
 
   computed: {
-    ...mapGetters('dataset', ['getUpdatedTime', 'getTimeRemaining']),
+    ...mapState('dataset', ['datasetInfo', 'analysisPlan']),
+    ...mapGetters('dataset', ['getDepositorSetupInfo', 'getUpdatedTime', 'getTimeRemaining']),
 
     isContinueDisabled: function () {
       return !this.steps[this.stepperPosition].completed;
